fix(select): handle model updates when options are manipulated
When an option that is currently selected, is removed or its value changes, the model
is set to null.
Likewise if an option is added or its value changes to the currently selected model,
this option is selected.
If an option is disabled, the model is set to null. If the model value changes to a
value that matches a disabled option, this option is selected (analogue to ngOptions)