fix(input[number]): validate the viewValue for min/max
This makes it consistent with all other input types, whose viewValue
is not modified via a parser.

BREAKING CHANGE:

Input[type=number] now validates the ngModel.$viewValue instead of
the ngModel.$modelValue, i.e. the value that is visible to the user
is validated. This makes it consistent with all other input types
for which the $viewValue is not modified via a parser. It also
makes more sense to the user, as the modelValue might be modified,
and a validation error would not necessarily reflect what the user
sees in the input.